Lesson-5
Несколько увлекся и залез чут дальше в middle.
Все работает.
