+++ package.json
@@ -25,13 +25,19 @@
         "@babel/preset-env": "^7.6.3",
         "@babel/preset-react": "^7.6.3",
         "babel-loader": "^8.0.6",
+        "connected-react-router": "^6.5.2",
         "css-loader": "^3.2.0",
+        "history": "^4.10.1",
         "material-ui": "^0.20.2",
         "package.json": "^2.0.1",
         "prop-types": "^15.7.2",
         "react": "^16.10.2",
+        "react-addons-update": "^15.6.2",
         "react-dom": "^16.10.2",
+        "react-redux": "^7.1.3",
         "react-router-dom": "^5.1.2",
+        "redux": "^4.0.4",
+        "redux-persist": "^6.0.0",
         "style-loader": "^1.0.0",
         "webpack": "^4.41.2",
         "webpack-cli": "^3.3.9",
