--- static_src/components/Profile.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-
-export default class Profile extends React.Component {
-    render() {
-        return (
-            <div className='profile'>
-                <h1>Профиль</h1>
-                <p>Lorem ipsum dolor sit amet consectetur, adipisicing elit.
-                        Aspernatur ratione doloremque nesciunt hic deserunt aut possimus,
-            ea, sed ex quo quia suscipit minus porro, officiis eligendi quos soluta eius magni.</p>
-            </div>
-        )
-    }
-}
-
