--- static_src/components/Router.jsx
@@ -1,21 +1,0 @@
-import React from 'react';
-import { Switch, Route } from 'react-router-dom';
-import Layout from './Layout.jsx';
-import Profile from './Profile.jsx';
-
-export default class Router extends React.Component {
-    render() {
-        return (
-            <Switch>
-                <Route exact path='/' component={Layout} />
-                <Route exact path='/chat/:chatId/'
-                    render={obj => {
-                        return <Layout chatId={Number(obj.match.params.chatId)} />
-                    }}
-                />
-                <Route exact path='/profile/' component={Profile} />
-
-            </Switch>
-        )
-    }
-}
