+++ static_src/middlewares/index.js
@@ -1,0 +1,5 @@
+import messageMiddleware from './messageMiddleware.js';
+
+export default [
+    messageMiddleware,
+];
