+++ static_src/reducers/messageReducer.js
@@ -1,0 +1,21 @@
+import update from 'react-addons-update';
+import { SEND_MESSAGE } from '../actions/messageActions.js';
+
+const initialStore = {
+    messages: {
+        1: { text: 'Привет!!!', sender: 'bot' },
+        2: { text: 'Как делишки?', sender: 'bot' },
+    },
+};
+
+export default function chatReducer(store = initialStore, action) {
+    switch (action.type) {
+        case SEND_MESSAGE: {
+            return update(store, {
+                messages: { $merge: { [action.messageId]: { text: action.text, sender: action.sender } } },
+            });
+        }
+        default:
+            return store;
+    }
+}
