--- static_src/script.js
@@ -1,3 +1,0 @@
-export const script = () => {
-    console.log('Это файл script.js');
-};
